requirements.txt
python-pptx
pandas
sentence-transformers
faiss-cpu
pysqlite3
openai==1.3.7
